requirements.txt
boto3==1.34.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
pytest==8.0.0
moto==5.0.0
